Creates SortButtons + ListInfo components
